+++ src/pytgpt/__init__.py
@@ -1,18 +1,11 @@
 from .utils import appdir
 import g4f
 
-__version__ = "0.3.5"
+__version__ = "0.3.6"
 __author__ = "Smartwa"
 __repo__ = "https://github.com/Simatwa/python-tgpt"
 
-tgpt_providers = [
-    "leo",
-    "openai",
-    "fakeopen",
-    "opengpt",
-    "koboldai",
-    "bard",
-]
+tgpt_providers = ["leo", "openai", "fakeopen", "opengpt", "koboldai", "bard", "phind"]
 
 gpt4free_providers = [
     provider.__name__ for provider in g4f.Provider.__providers__  # if provider.working